Modification de Exemplaire et Oeuvre delete
les delete disent à chaque fois qu'il y a des enregistrements a supprimer

+++ Oeuvre_delete.php
@@ -5,10 +5,20 @@ include("connexion_bdd.php");
 if(isset($_GET["id"]) AND is_numeric($_GET["id"]))
 {
 	$id=htmlentities($_GET['id']);
-	$ma_requete_SQL="DELETE FROM OEUVRE WHERE noOeuvre = ".$id.";";
-	$bdd->exec($ma_requete_SQL);
+	$check_requette="SELECT *FROM EXEMPLAIRE WHERE noOeuvre =".$id.";";
+	$result=$bdd->query($check_requette);
+	$result=$result->fetchAll();
+	if(empty($result)){
 
-	header("Location: Oeuvre_show.php");
+		$ma_requete_SQL="DELETE FROM OEUVRE WHERE noOeuvre = ".$id.";";
+		$bdd->exec($ma_requete_SQL);
+		header("Location: Oeuvre_show.php");
+
+	}else{
+		echo "attention, il reste des exemplaires a supprimer <br>";
+		echo "<a href='Exemplaire_show.php'>exemplaires</a><br>";
+		echo "<a href='Oeuvre_show.php'>revenir en arrière</a><br>";
+	}
 }
 
 ?>
